Report unreadable include files
This happens a lot when ini_set('include_path', ...) is called within
the PHP file.  We don't want to track calls to ini_set() since it is
very difficult to tell what 'include_path' is being set to, so best to
provide a -i option so that php-initialized can find include files.