Provide a -i option to provide include path(s)
This can be used to provide a list of directories in the project, and
also a list of library directories.  Unfortunately, this can lead to
php-initialized reporting all uninitialised variables in the library,
i.e. not your code.  Perhaps provide an -l option for libraries and
don't report uninitialised variables found within the library files.